Split boxed primitive formatting helpers
